move specifier module into summary module
